requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15,<0.16
pillow>=8.2.0,<8.3.0
celery>=5.2.0,<5.3
requests>=2.25,<3.0
chardet>=5.2.0,<5.3
django-import-export>=3.3.7,<3.4
